requirements.txt
asgiref==3.8.1
Django==5.0.9
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
mssql-django==1.4
pillow==10.4.0
PyJWT==2.9.0
pyodbc==5.2.0
python-dotenv==1.0.1
pytz==2024.2
sqlparse==0.5.1
tzdata==2024.2
